+++ rust/crd/src/lib.rs
@@ -1,8 +1,13 @@
+use k8s_openapi::apimachinery::pkg::apis::meta::v1::Condition;
 use kube::CustomResource;
 use schemars::JsonSchema;
+use semver::Version;
 use serde::{Deserialize, Serialize};
 use stackable_operator::product_config_utils::{ConfigError, Configuration};
 use stackable_operator::role_utils::Role;
+use stackable_operator::status::{Conditions, Status, Versioned};
+use stackable_operator::versioning::{ProductVersion, Versioning, VersioningState};
+use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use strum_macros::Display;
 use strum_macros::EnumIter;
@@ -32,6 +37,15 @@ pub struct NifiSpec {
     pub zookeeper_reference: stackable_zookeeper_crd::util::ZookeeperReference,
 }
 
+impl Status<NifiStatus> for NifiCluster {
+    fn status(&self) -> &Option<NifiStatus> {
+        &self.status
+    }
+    fn status_mut(&mut self) -> &mut Option<NifiStatus> {
+        &mut self.status
+    }
+}
+
 #[allow(non_camel_case_types)]
 #[derive(
     Clone,
@@ -49,10 +63,73 @@ pub enum NifiVersion {
     #[serde(rename = "1.13.2")]
     #[strum(serialize = "1.13.2")]
     v1_13_2,
+
+    // TODO: NiFi 1.14 does not work with this operator yet <https://github.com/stackabletech/nifi-operator/issues/82>
+    //    Therefore we skip using it in the CRD for now. Should be uncommented as soon as #82 is fixed.
+    //#[serde(rename = "1.14.0")]
+    #[serde(skip)]
+    #[strum(serialize = "1.14.0")]
+    v1_14_0,
+}
+
+impl Versioning for NifiVersion {
+    fn versioning_state(&self, other: &Self) -> VersioningState {
+        let from_version = match Version::parse(&self.to_string()) {
+            Ok(v) => v,
+            Err(e) => {
+                return VersioningState::Invalid(format!(
+                    "Could not parse [{}] to SemVer: {}",
+                    self.to_string(),
+                    e.to_string()
+                ))
+            }
+        };
+
+        let to_version = match Version::parse(&other.to_string()) {
+            Ok(v) => v,
+            Err(e) => {
+                return VersioningState::Invalid(format!(
+                    "Could not parse [{}] to SemVer: {}",
+                    other.to_string(),
+                    e.to_string()
+                ))
+            }
+        };
+
+        match to_version.cmp(&from_version) {
+            Ordering::Greater => VersioningState::ValidUpgrade,
+            Ordering::Less => VersioningState::ValidDowngrade,
+            Ordering::Equal => VersioningState::NoOp,
+        }
+    }
 }
 
 #[derive(Clone, Debug, Default, Deserialize, JsonSchema, Serialize)]
-pub struct NifiStatus {}
+pub struct NifiStatus {
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    #[schemars(schema_with = "stackable_operator::conditions::schema")]
+    pub conditions: Vec<Condition>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub version: Option<ProductVersion<NifiVersion>>,
+}
+
+impl Versioned<NifiVersion> for NifiStatus {
+    fn version(&self) -> &Option<ProductVersion<NifiVersion>> {
+        &self.version
+    }
+    fn version_mut(&mut self) -> &mut Option<ProductVersion<NifiVersion>> {
+        &mut self.version
+    }
+}
+
+impl Conditions for NifiStatus {
+    fn conditions(&self) -> &[Condition] {
+        self.conditions.as_slice()
+    }
+    fn conditions_mut(&mut self) -> &mut Vec<Condition> {
+        &mut self.conditions
+    }
+}
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -120,4 +197,34 @@ impl Configuration for NifiConfig {
 pub enum NifiRole {
     #[strum(serialize = "node")]
     Node,
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::NifiVersion;
+    use stackable_operator::versioning::{Versioning, VersioningState};
+    use std::str::FromStr;
+
+    #[test]
+    fn test_zookeeper_version_versioning() {
+        assert_eq!(
+            NifiVersion::v1_13_2.versioning_state(&NifiVersion::v1_14_0),
+            VersioningState::ValidUpgrade
+        );
+        assert_eq!(
+            NifiVersion::v1_14_0.versioning_state(&NifiVersion::v1_13_2),
+            VersioningState::ValidDowngrade
+        );
+        assert_eq!(
+            NifiVersion::v1_13_2.versioning_state(&NifiVersion::v1_13_2),
+            VersioningState::NoOp
+        );
+    }
+
+    #[test]
+    fn test_version_conversion() {
+        NifiVersion::from_str("1.13.2").unwrap();
+        NifiVersion::from_str("1.14.0").unwrap();
+        NifiVersion::from_str("1.2.3").unwrap_err();
+    }
 }